refactor: move MyForm from lib to Components
